requirements.txt
tensorflow
tensorflow.keras
torch
torchvision
pillow
numpy
matplotlib
scipy
pandas
scikit-learn
opencv-python
tqdm
requests
